requirements.txt
google-genai
python-dotenv==1.1.1
pillow==11.3.0
